+++ NEWS.md
@@ -5,6 +5,7 @@
 - New feature: `onevent()` accepts a `properties` parameter that allows the user to retrieve additional properties that are not whitelisted by default; fixes #159
 - Fix bug: shinyjs functions used inside a module with a `selector` argument instead of an `id` argument didn't work
 - Add `asis` parameter to any function that takes an ID. When `asis=TRUE`, the ID will not be namespaced when inside a module (#118)
+- The `includeShinyjs` parameter in `runcodeUI()` is now deprecated
 
 # shinyjs 1.0